fix parser: cast to int64_t before multiply to avoid overflow (CodeQL)
ty->size is int64_t, but (i + 1) * wchar_size was computed in 32-bit
int before widening. Cast (i + 1) to int64_t so the multiplication
happens in 64-bit, preventing overflow with large wide-string arrays.

Fixes https://github.com/rurban/rcc/security/code-scanning/1

+++ src/parser.c
@@ -2474,7 +2474,7 @@ static Token *global_init_one(Token *tok, LVar *var, Type *ty, int offset) {
         }
         // If array size is 0 (incomplete), set it
         if (ty->size == 0)
-            ty->size = (i + 1) * wchar_size;
+            ty->size = (int64_t)(i + 1) * wchar_size;
         return tok->next;
     }
